+++ lib/rules/require-extension.js
@@ -1,77 +1,32 @@
 /**
  * @fileoverview Restrict file extensions that may be required
  * @author Scott Andrews
+ * @deprecated
  */
 'use strict';
-
-var path = require('path');
-
-// ------------------------------------------------------------------------------
-// Constants
-// ------------------------------------------------------------------------------
-
-var DEFAULTS = {
-  extensions: ['.jsx']
-};
-
-var PKG_REGEX = /^[^\.]((?!\/).)*$/;
 
 // ------------------------------------------------------------------------------
 // Rule Definition
 // ------------------------------------------------------------------------------
 
+var util = require('util');
+var jsxRequireExtension = require('./jsx-require-extension');
+var isWarnedForDeprecation = false;
+
 module.exports = function(context) {
-
-  function isPackage(id) {
-    return PKG_REGEX.test(id);
-  }
-
-  function isRequire(expression) {
-    return expression.callee.name === 'require';
-  }
-
-  function getId(expression) {
-    return expression.arguments[0] && expression.arguments[0].value;
-  }
-
-  function getExtension(id) {
-    return path.extname(id || '');
-  }
-
-  function getExtensionsConfig() {
-    return context.options[0] && context.options[0].extensions || DEFAULTS.extensions;
-  }
-
-  var forbiddenExtensions = getExtensionsConfig().reduce(function (extensions, extension) {
-    extensions[extension] = true;
-    return extensions;
-  }, Object.create(null));
-
-  function isForbiddenExtension(ext) {
-    return ext in forbiddenExtensions;
-  }
-
-  // --------------------------------------------------------------------------
-  // Public
-  // --------------------------------------------------------------------------
-
-  return {
-
-    CallExpression: function(node) {
-      if (isRequire(node)) {
-        var id = getId(node);
-        var ext = getExtension(id);
-        if (!isPackage(id) && isForbiddenExtension(ext)) {
-          context.report({
-            node: node,
-            message: 'Unable to require module with extension \'' + ext + '\''
-          });
-        }
+  return util._extend(jsxRequireExtension(context), {
+    Program: function() {
+      if (isWarnedForDeprecation || /\=-(f|-format)=/.test(process.argv.join('='))) {
+        return;
       }
+
+      /* eslint-disable no-console */
+      console.log('The react/require-extension rule is deprecated. Please ' +
+                  'use the react/jsx-require-extension rule instead.');
+      /* eslint-enable no-console */
+      isWarnedForDeprecation = true;
     }
-
-  };
-
+  });
 };
 
 module.exports.schema = [{
